ChangeDetail has its own script file
