+++ index.js
@@ -1,5 +1,6 @@
 
 module.exports = {
 	'async': require('./async.js'),
-	'sync': require('./sync.js')
+	'sync': require('./sync.js'),
+	'utils': require('./utils')
 };
