+++ utils/change-detail.js
@@ -1,0 +1,15 @@
+
+((module) => {
+    'use strict';
+
+    function ChangeDetail(type, name, prevmtime, currmtime) {
+		this.type = type;
+		this.name = name;
+		this.prevmtime = prevmtime;
+		this.currmtime = currmtime;
+		freeze(this);
+	}
+
+    module.exports = class extends ChangeDetail {};
+
+})(module);
