+++ utils/index.js
@@ -1,0 +1,5 @@
+
+module.exports = {
+    'ChangeDetail': require('./change-detail.js'),
+    '__proto__': null
+}
